1. 完成出库模块. 2. 将阿里图标库设置到本地,追加对应过滤器排除的扩展名. 3. 货品删除时,需要不存在剩余库存. 4. 删除入库单时,若已审核通过则不能删除.在需要删除时,一并删除审核记录. 5. 将删除的响应图改为小写. 6. 入库审核若已通过,则无法再对审核进行修改. 7. 修正时间显示,修改系统初始页面. 8. 修正Controller层的script没有结束标签导致的无法跳转与提示BUG.

+++ src/com/controll/GoodsController.java
@@ -111,7 +111,7 @@ public class GoodsController {
             application.setAttribute("goods",GoodsController.initializationGoods(ms));
             response.setHeader("refresh", "0;URL=goodsList");
         } else {
-            response.getWriter().print("<script>alert('修改失败');window.history.go(-1);");
+            response.getWriter().print("<script>alert('修改失败');window.history.go(-1);</script>");
         }
     }
     @RequestMapping("/GoodsDel")
@@ -122,7 +122,11 @@ public class GoodsController {
             response.getWriter().print("<script>alert('非法访问没有传值');window.location='goodsList'</script>");
             return;
         }
-        //todo 待库存表关联
+        ms.setSql("select * from inventory where gid=?").set(Integer.parseInt(gid));
+        if (ms.runList().size()>0){
+            response.getWriter().print("<script>alert('删除失败,仓库已存在货品');window.location='goodsList'</script>");
+            return;
+        }
         //----------------------------------------删除-------------------------------------------------------------------
         ms.setSql("DELETE FROM goods where gid=?").set(Integer.parseInt(gid));
         if (ms.run() > 0) {
